requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
psycopg[binary]==3.3.2
email-validator==2.3.0
httpx==0.27.0
aiofiles==23.2.0
